ai-service/pyproject.toml
[project]
name = "ioffer-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.6.2",
    "autogen-ext[openai]>=0.6.2",
    "autogen-ext[web-surfer]>=0.6.2",
    "chainlit>=2.6.0",
    "loguru>=0.7.3",
    "pymongo>=4.13.2",
    "pymysql>=1.1.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pandas>=2.0.0",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.0",
    "pytest>=8.0.0",
    "pydantic-settings>=2.0.0",
    "mysql-connector-python>=9.3.0",
    "pandas>=2.0.0",
    "pdf2image>=1.17.0",
    "joblib>=1.5.1",
    "scikit-learn>=1.7.1",
    "xgboost>=3.0.3",
    "google-cloud-storage>=3.2.0",
    "google-genai>=1.29.0",
    "matplotlib>=3.10.5",
    "seaborn>=0.13.2",
    "pyjwt>=2.10.1",
]
